requirements.txt
Django==5.2.4
djangorestframework==3.16.0
gunicorn>=21.2.0
pillow==11.3.0